shorten comments, revert changes waitForSettledTopmostIndex and implement it different way
